+++ test/astranaut_macro_SUITE_data/macro_export_local_options_warning_test.erl
@@ -5,7 +5,7 @@
 -baseline(yep).
 
 -export_macro({exported/0,
-               [{extra_functions, [{missing, 0}]},
+               [{closure_roots, [{missing, 0}]},
                 {internal_function, true}]}).
 
 exported() ->