+++ test/astranaut_macro_SUITE_data/macro_local_internal_option_warning_test.erl
@@ -1,0 +1,10 @@
+-module(macro_local_internal_option_warning_test).
+
+-include("macro.hrl").
+
+-baseline(yep).
+
+-local_macro({entry/0, [{internal_function, true}]}).
+
+entry() ->
+    {atom, 0, ok}.