+++ test/astranaut_macro_SUITE_data/macro_pass_closure_roots_test.erl
@@ -1,14 +1,14 @@
 %%%-------------------------------------------------------------------
-%%% extra_functions can add helpers missed by static call scanning.
+%%% closure_roots can add helpers missed by static call scanning.
 %%%-------------------------------------------------------------------
--module(macro_pass_extra_functions_test).
+-module(macro_pass_closure_roots_test).
 
 -include("quote.hrl").
 -include("macro.hrl").
 
 -export([value/0]).
 
--local_macro({[entry/1], [{extra_functions, [{hidden_helper, 1}]}]}).
+-local_macro({[entry/1], [{closure_roots, [{hidden_helper, 1}]}]}).
 
 value() ->
     entry(ok).